Refactor db::GetIndex into DbCustomIndex
This is a part of #3.

This patch renames struct to show the purpose - it is a struct which
holds information about a custom index. This patch moves the struct to
the lib.rs as a preparation for exporting API for integration tests.
